[AscendNPU IR] Partially reconstructed the jit_npu
